+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.1
+- Moved the Bar class to bar_chart.dart
+- Fixed a bug in bar_chart.dart where BarChart required the bar package instead of corechart.
+
 ## 0.1.0
 - Add: DataView class, formatters, static methods
 - Add: Examples for format class